requirements.txt
ccxt
openai
pandas
schedule
python-dotenv
requests
urllib3